+++ meiqiasdk/src/main/java/com/meiqia/meiqiasdk/imageloader/MQImage.java
@@ -1,5 +1,6 @@
 package com.meiqia.meiqiasdk.imageloader;
 
+import android.app.Activity;
 import android.content.Context;
 import android.support.annotation.DrawableRes;
 import android.widget.ImageView;
@@ -20,7 +21,7 @@ public class MQImage {
             synchronized (MQImage.class) {
                 if (sImageLoader == null) {
                     if (isClassExists("com.bumptech.glide.Glide")) {
-                        sImageLoader = new MQGlideImageloader();
+                        sImageLoader = new MQGlideImageLoader();
                     } else if (isClassExists("com.squareup.picasso.Picasso")) {
                         sImageLoader = new MQPicassoImageLoader();
                     } else if (isClassExists("com.nostra13.universalimageloader.core.ImageLoader")) {
@@ -45,8 +46,8 @@ public class MQImage {
         }
     }
 
-    public static void displayImage(final ImageView imageView, String path, @DrawableRes int loadingResId, @DrawableRes int failResId, int width, int height, final MQImageLoader.MQDisplayImageListener delegate) {
-        getImageLoader().displayImage(imageView, path, loadingResId, failResId, width, height, delegate);
+    public static void displayImage(Activity activity, ImageView imageView, String path, @DrawableRes int loadingResId, @DrawableRes int failResId, int width, int height, final MQImageLoader.MQDisplayImageListener delegate) {
+        getImageLoader().displayImage(activity, imageView, path, loadingResId, failResId, width, height, delegate);
     }
 
     public static void downloadImage(Context context, String path, final MQImageLoader.MQDownloadImageListener delegate) {